refactor(image-recognition): improve thread handling
