+++ app/src/main/java/com/comp90018/contexttunes/data/viewModel/ImageViewModel.java
@@ -6,7 +6,7 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-import com.comp90018.contexttunes.services.ImageLabels;
+import com.comp90018.contexttunes.domain.ImageLabels;
 
 public class ImageViewModel extends ViewModel {
     private final MutableLiveData<Bitmap> capturedImage = new MutableLiveData<>(); // this  holds a Bitmap that can change over time
@@ -29,11 +29,19 @@ public class ImageViewModel extends ViewModel {
     }
 
     /**
-     * This method stores the labels for the image in a MutableLiveData object.
+     * This method stores the labels for the image in a MutableLiveData object. (main thread)
      * @param labels
      */
     public void setImageLabels(ImageLabels labels) {
         imageLabels.setValue(labels);
+    }
+
+    /**
+     * Store labels from a background thread (used by ImageAnalyser).
+     * @param labels
+     */
+    public void postImageLabels(ImageLabels labels) {
+        imageLabels.postValue(labels);
     }
 
     /**
