+++ local.defaults.properties
@@ -1,2 +1,3 @@
 PLACES_API_KEY=DEFAULT_API_KEY
 OPENWEATHER_API_KEY=DEFAULT_OPENWEATHER_API_KEY
+AWS_MODEL_KEY=DEFAULT_AWS_MODEL_KEY
